ENGR00217015 Music: Thumbnail of each song show Gallery's picture
MediaProvider will try to fall back to the image in the same folder
of music, if MetadataRetriver failed to extract the AlbumArt image.
MediaProvider will bypass the albumArt image search if the music located
in the root external storage. For those mp3 file in the root dir of /extsd,
need to add the extsd and udisk dir checking

Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>